// dashboards-observability/common/types/explorer.ts
export interface IField {
  name: string;
  type: string;
  label?: string;
}

export interface ExplorerFields {
  availableFields: IField[];
  queriedFields: IField[];
  selectedFields: IField[];
  unselectedFields: IField[];
}

export interface ITabQueryResults {
  [tabId: string]: any;
}

export interface ITabQueries {
  [tabId: string]: IQuery;
}

export interface IQuery {
  [RAW_QUERY]: string;
  [FINAL_QUERY]: string;
  [INDEX]: string;
  [SELECTED_DATE_RANGE]: string[];
  [SELECTED_TIMESTAMP]: string;
}
// ...
export interface ExplorerData {
  explorerData?: ExplorerDataType;
  explorerFields?: IExplorerFields;
  query?: Query;
  http?: HttpSetup;
  pplService?: PPLService;
}

export interface IVisualizationContainerPropsData {
  appData?: { fromApp: boolean };
  rawVizData?: any;
  query?: IQuery;
  indexFields?: IField[];
  userConfigs?: any;
  defaultAxes?: {
    xaxis: IField[];
    yaxis: IField[];
  };
  explorer?: ExplorerData;
}

export interface IVisualizationContainerPropsVis {
  vis: IVisualizationTypeDefination;
}
// ...
export interface IVisualizationTypeDefination {
  name: string;
  type: string;
  id: string;
  label: string;
  fulllabel: string;
  category: string;
  icon: React.ReactNode;
  editorconfig: {
    panelTabs: IConfigPanelTab;
  };
  visconfig: {
    layout: Partial<Plotly.Layout>;
    config: Partial<Plotly.Config>;
  };
  component: React.ReactNode;
}

export interface IVisualizationContainerProps {
  data: IVisualizationContainerPropsData;
  vis: IVisualizationContainerPropsVis;
}